move GCS utils to a separate package
importing bootstramp/cmd from transit/cmd caused the transit binary to
require CGO and relic
